requirements.txt
feedparser==6.0.11
requests==2.32.3
openai==1.30.5
anthropic==0.28.0
python-dotenv==1.0.1
jinja2==3.1.4
